refactor transform value cast
